started refactor of sections/questions
