+++ app/views/org_admin/sections/_edit.html.erb
@@ -5,69 +5,11 @@
 </div>
 <div class="row">
   <div class="col-md-12">
-    <%= form_for(
-        section,
-        url: org_admin_template_phase_section_path(template_id: template.id, phase_id: phase.id, id: section),
-        namespace: section.id,
-        html: { method: :put }) do |f| %>
-      <div class="form-group col-md-10">
-        <%= f.label(:title, _('Title') ,class: "control-label") %>
-        <%= f.text_field(:title, { class: "form-control", 'aria-required': false, placeholder: _('Enter a title for the section'), 'data-toggle': 'tooltip', title: _('Enter a title for the section')} ) %>
-      </div>
-
-      <div class="form-group">
-        <div class="col-md-10">
-          <%= f.label(:number, _('Order of display'), class: "control-label") %>
-        </div>
-        <div class="col-md-4">
-          <%= f.number_field(:number, in: 1..15, class: "form-control", 'aria-required': false, 'data-toggle': 'tooltip', title: _('This allows you to order sections.')) %>
-        </div>
-      </div>
-
-      <div class="form-group col-md-10">
-        <%= f.label(:description, _('Description'), class: "control-label") %>
-        <div data-toggle="tooltip" title="<%= _("Enter a basic description. This could be a summary of what is covered in the section or instructions on how to answer. This text will be displayed in the coloured banner once a section is opened to edit.") %>">
-          <%= f.text_area(:description, class: "section tinymce form-control") %>
-        </div>
-      </div>
-
-      <div class="form-group col-md-10">
-        <div class="pull-right">
-          <%= f.button(_('Save'), class: 'btn btn-default', type: "submit") %>
-          <%= link_to _('Delete'), org_admin_template_phase_section_path(template_id: template.id, phase_id: phase.id, id: section), 'data-method': 'delete', rel: 'nofollow',
-                'data-confirm': _("You are about to delete '%{section_title}'. This will affect questions linked to this section. Are you sure?") % { :section_title => section.title }, class: 'btn btn-default', role:'button' %>
-          <%= link_to _('Cancel'), edit_org_admin_template_phase_path(template, phase), class: 'btn btn-default', role: 'button' %>
-        </div>
-      </div>
-    <% end %>
+    <%= render partial: 'org_admin/sections/form',
+               locals: local_assigns.merge({ method: :put, url: org_admin_template_phase_section_path(template_id: template.id, phase_id: section.phase.id, id: section) }) %>
   </div>
   <div class="col-md-12">
-    <h3><%= _('Questions') %></h3>
-   <!--
-    <%# questions = section.questions.order('number') %>
-    <%# if questions.present? %>
-        <%# questions.each do |question| %>
-          <hr />
-          <div class="question_show" id="<%#= "question_show#{question.id}" %>">
-            <%#= render partial: 'org_admin/questions/show', locals: {template: template, question: question, current_tab: current_tab, edit: edit, original_org: template.base_org} %>
-          </div>
-          <div class="question_edit" id="<%#= "question_edit#{question.id}" %>" style="display: none;">
-            <%#= render partial: 'org_admin/questions/edit', locals: {template: template, question: question, current_tab: current_tab} %>
-          </div>
-          <%#= raw("<hr />") if questions.last.id == question.id %>
-        <%# end %>
-    <%# end %>
-  </div-->
+    <%= render partial: 'org_admin/questions/index', 
+               locals: local_assigns.merge({ editing: true }) %>
+  </div>
 </div>
-<%# if edit %>
-  <!--div class="row">
-    <div class="col-md-12">
-      <div class="pull-right">
-        <%#= link_to(_('Add Question'), '#', { class: 'btn btn-default question_new_link', role: "button" }) %>
-      </div>
-      <div class="question_new" style="display: none;">
-        <%#= render partial: 'org_admin/questions/new', locals: { template: template, phase: phase, section: section, current_tab: current_tab } %>
-      </div>
-    </div>
-  </div-->
-<%# end %>
