+++ lib/assets/javascripts/application.js
@@ -20,7 +20,14 @@ import './views/guidance_groups/admin_edit';
 import './views/guidance_groups/admin_new';
 import './views/guidances/new_edit';
 import './views/notes/index';
-import './views/org_admin/phases/edit';
+import './views/org_admin/phases/new_edit';
+import './views/org_admin/questions/new_edit';
+import './views/org_admin/questions/index';
+import './views/org_admin/questions/show';
+import './views/org_admin/questions/sharedEventHandlers';
+import './views/org_admin/question_options/index';
+import './views/org_admin/sections/index';
+import './views/org_admin/sections/new_edit';
 import './views/org_admin/templates/edit';
 import './views/org_admin/templates/index';
 import './views/org_admin/templates/new';
@@ -33,13 +40,6 @@ import './views/plans/index';
 import './views/plans/new';
 import './views/plans/share';
 import './views/roles/edit';
-import './views/sections/edit';
-import './views/sections/index';
-import './views/sections/new';
-import './views/questions/new_edit';
-import './views/questions/index';
-import './views/questions/show';
-import './views/question_options/index';
 import './views/shared/create_account_form';
 import './views/shared/my_org';
 import './views/shared/sign_in_form';
