+++ lib/assets/javascripts/views/org_admin/phases/new_edit.js
@@ -1,12 +1,13 @@
 import 'bootstrap-sass/assets/javascripts/bootstrap/collapse';
 import expandCollapseAll from '../../../utils/expandCollapseAll';
 import { Tinymce } from '../../../utils/tinymce';
+import ariatiseForm from '../../../utils/ariatiseForm';
 
 $(() => {
   // Attach handlers for the expand/collapse all accordions
   expandCollapseAll();
   Tinymce.init({ selector: '.phase' });
-
+  ariatiseForm({ selector: '.phase_form' });
   const parentSelector = '#sections_accordion';
   $(parentSelector).on('ajax:before', (e) => {
     const panelBody = $(e.target).parent().find('.panel-body');
@@ -16,7 +17,8 @@ $(() => {
     const panelBody = $(e.target).parent().find('.panel-body');
     panelBody.attr('data-loaded', 'true');
     panelBody.html(data);
-    Tinymce.init({ selector: '.section' });
+    Tinymce.init({ selector: '.section, .question' });
+    ariatiseForm({ selector: '.section_form, .question_form' });
   });
   $(parentSelector).on('ajax:error', 'a[data-remote="true"]', () => {
     // TODO something generic for every error
