+++ lib/assets/javascripts/views/org_admin/questions/sharedEventHandlers.js
