+++ lib/assets/javascripts/views/org_admin/questions/show.js
