+++ lib/assets/javascripts/views/org_admin/sections/index.js
