--- lib/assets/javascripts/views/sections/edit.js
@@ -1,5 +1,0 @@
-import { Tinymce } from '../../utils/tinymce';
-
-$(() => {
-  Tinymce.init({ selector: '.section' });
-});
