--- lib/assets/javascripts/views/sections/new.js
@@ -1,7 +1,0 @@
-$(() => {
-  $('.section_new_cancel').on('click', (e) => {
-    const sectionNew = $(e.target).closest('.section_new');
-    sectionNew.hide();
-    sectionNew.closest('.row').find('.section_new_link').show();
-  });
-});
